+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/cache/FilterSupportCacheImpl.java
@@ -3,52 +3,96 @@ package com.dwarfeng.fdr.impl.cache;
 import com.dwarfeng.fdr.sdk.bean.entity.FastJsonFilterSupport;
 import com.dwarfeng.fdr.stack.bean.entity.FilterSupport;
 import com.dwarfeng.fdr.stack.cache.FilterSupportCache;
-import com.dwarfeng.subgrade.impl.cache.RedisBaseCache;
+import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.CacheException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Repository
 public class FilterSupportCacheImpl implements FilterSupportCache {
 
-    @Autowired
-    private RedisBaseCache<StringIdKey, FilterSupport, FastJsonFilterSupport> baseCache;
+    private final RedisBatchBaseCache<StringIdKey, FilterSupport, FastJsonFilterSupport>
+            filterSupportBatchBaseDelegate;
+
+    public FilterSupportCacheImpl(
+            RedisBatchBaseCache<StringIdKey, FilterSupport, FastJsonFilterSupport> filterSupportBatchBaseDelegate
+    ) {
+        this.filterSupportBatchBaseDelegate = filterSupportBatchBaseDelegate;
+    }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean exists(StringIdKey key) throws CacheException {
-        return baseCache.exists(key);
+        return filterSupportBatchBaseDelegate.exists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public FilterSupport get(StringIdKey key) throws CacheException {
-        return baseCache.get(key);
+        return filterSupportBatchBaseDelegate.get(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void push(FilterSupport value, long timeout) throws CacheException {
-        baseCache.push(value, timeout);
+        filterSupportBatchBaseDelegate.push(value, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void delete(StringIdKey key) throws CacheException {
-        baseCache.delete(key);
+        filterSupportBatchBaseDelegate.delete(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void clear() throws CacheException {
-        baseCache.clear();
+        filterSupportBatchBaseDelegate.clear();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return filterSupportBatchBaseDelegate.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return filterSupportBatchBaseDelegate.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<FilterSupport> batchGet(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return filterSupportBatchBaseDelegate.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchPush(@SkipRecord List<FilterSupport> entities, long timeout) throws CacheException {
+        filterSupportBatchBaseDelegate.batchPush(entities, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        filterSupportBatchBaseDelegate.batchDelete(keys);
     }
 }