+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/cache/MapperSupportCacheImpl.java
@@ -3,52 +3,96 @@ package com.dwarfeng.fdr.impl.cache;
 import com.dwarfeng.fdr.sdk.bean.entity.FastJsonMapperSupport;
 import com.dwarfeng.fdr.stack.bean.entity.MapperSupport;
 import com.dwarfeng.fdr.stack.cache.MapperSupportCache;
-import com.dwarfeng.subgrade.impl.cache.RedisBaseCache;
+import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.CacheException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Repository
 public class MapperSupportCacheImpl implements MapperSupportCache {
 
-    @Autowired
-    private RedisBaseCache<StringIdKey, MapperSupport, FastJsonMapperSupport> baseCache;
+    private final RedisBatchBaseCache<StringIdKey, MapperSupport, FastJsonMapperSupport>
+            mapperSupportBatchBaseDelegate;
+
+    public MapperSupportCacheImpl(
+            RedisBatchBaseCache<StringIdKey, MapperSupport, FastJsonMapperSupport> mapperSupportBatchBaseDelegate
+    ) {
+        this.mapperSupportBatchBaseDelegate = mapperSupportBatchBaseDelegate;
+    }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean exists(StringIdKey key) throws CacheException {
-        return baseCache.exists(key);
+        return mapperSupportBatchBaseDelegate.exists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public MapperSupport get(StringIdKey key) throws CacheException {
-        return baseCache.get(key);
+        return mapperSupportBatchBaseDelegate.get(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void push(MapperSupport value, long timeout) throws CacheException {
-        baseCache.push(value, timeout);
+        mapperSupportBatchBaseDelegate.push(value, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void delete(StringIdKey key) throws CacheException {
-        baseCache.delete(key);
+        mapperSupportBatchBaseDelegate.delete(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void clear() throws CacheException {
-        baseCache.clear();
+        mapperSupportBatchBaseDelegate.clear();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return mapperSupportBatchBaseDelegate.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return mapperSupportBatchBaseDelegate.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<MapperSupport> batchGet(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return mapperSupportBatchBaseDelegate.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchPush(@SkipRecord List<MapperSupport> entities, long timeout) throws CacheException {
+        mapperSupportBatchBaseDelegate.batchPush(entities, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        mapperSupportBatchBaseDelegate.batchDelete(keys);
     }
 }