+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/cache/TriggerSupportCacheImpl.java
@@ -3,52 +3,96 @@ package com.dwarfeng.fdr.impl.cache;
 import com.dwarfeng.fdr.sdk.bean.entity.FastJsonTriggerSupport;
 import com.dwarfeng.fdr.stack.bean.entity.TriggerSupport;
 import com.dwarfeng.fdr.stack.cache.TriggerSupportCache;
-import com.dwarfeng.subgrade.impl.cache.RedisBaseCache;
+import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.CacheException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Repository
 public class TriggerSupportCacheImpl implements TriggerSupportCache {
 
-    @Autowired
-    private RedisBaseCache<StringIdKey, TriggerSupport, FastJsonTriggerSupport> baseCache;
+    private final RedisBatchBaseCache<StringIdKey, TriggerSupport, FastJsonTriggerSupport>
+            triggerSupportBatchBaseDelegate;
+
+    public TriggerSupportCacheImpl(
+            RedisBatchBaseCache<StringIdKey, TriggerSupport, FastJsonTriggerSupport> triggerSupportBatchBaseDelegate
+    ) {
+        this.triggerSupportBatchBaseDelegate = triggerSupportBatchBaseDelegate;
+    }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean exists(StringIdKey key) throws CacheException {
-        return baseCache.exists(key);
+        return triggerSupportBatchBaseDelegate.exists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public TriggerSupport get(StringIdKey key) throws CacheException {
-        return baseCache.get(key);
+        return triggerSupportBatchBaseDelegate.get(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void push(TriggerSupport value, long timeout) throws CacheException {
-        baseCache.push(value, timeout);
+        triggerSupportBatchBaseDelegate.push(value, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void delete(StringIdKey key) throws CacheException {
-        baseCache.delete(key);
+        triggerSupportBatchBaseDelegate.delete(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void clear() throws CacheException {
-        baseCache.clear();
+        triggerSupportBatchBaseDelegate.clear();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return triggerSupportBatchBaseDelegate.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return triggerSupportBatchBaseDelegate.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<TriggerSupport> batchGet(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return triggerSupportBatchBaseDelegate.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchPush(@SkipRecord List<TriggerSupport> entities, long timeout) throws CacheException {
+        triggerSupportBatchBaseDelegate.batchPush(entities, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        triggerSupportBatchBaseDelegate.batchDelete(keys);
     }
 }