+++ fdr-stack/src/main/java/com/dwarfeng/fdr/stack/cache/FilterSupportCache.java
@@ -2,7 +2,7 @@ package com.dwarfeng.fdr.stack.cache;
 
 import com.dwarfeng.fdr.stack.bean.entity.FilterSupport;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
-import com.dwarfeng.subgrade.stack.cache.BaseCache;
+import com.dwarfeng.subgrade.stack.cache.BatchBaseCache;
 
 /**
  * 过滤器支持缓存。
@@ -10,5 +10,5 @@ import com.dwarfeng.subgrade.stack.cache.BaseCache;
  * @author DwArFeng
  * @since 0.0.1-alpha
  */
-public interface FilterSupportCache extends BaseCache<StringIdKey, FilterSupport> {
+public interface FilterSupportCache extends BatchBaseCache<StringIdKey, FilterSupport> {
 }