+++ fdr-stack/src/main/java/com/dwarfeng/fdr/stack/cache/MapperSupportCache.java
@@ -2,7 +2,7 @@ package com.dwarfeng.fdr.stack.cache;
 
 import com.dwarfeng.fdr.stack.bean.entity.MapperSupport;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
-import com.dwarfeng.subgrade.stack.cache.BaseCache;
+import com.dwarfeng.subgrade.stack.cache.BatchBaseCache;
 
 /**
  * 映射器支持缓存。
@@ -10,5 +10,5 @@ import com.dwarfeng.subgrade.stack.cache.BaseCache;
  * @author DwArFeng
  * @since 1.6.0
  */
-public interface MapperSupportCache extends BaseCache<StringIdKey, MapperSupport> {
+public interface MapperSupportCache extends BatchBaseCache<StringIdKey, MapperSupport> {
 }