+++ fdr-stack/src/main/java/com/dwarfeng/fdr/stack/cache/TriggerSupportCache.java
@@ -2,7 +2,7 @@ package com.dwarfeng.fdr.stack.cache;
 
 import com.dwarfeng.fdr.stack.bean.entity.TriggerSupport;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
-import com.dwarfeng.subgrade.stack.cache.BaseCache;
+import com.dwarfeng.subgrade.stack.cache.BatchBaseCache;
 
 /**
  * 触发器支持缓存。
@@ -10,5 +10,5 @@ import com.dwarfeng.subgrade.stack.cache.BaseCache;
  * @author DwArFeng
  * @since 0.0.1-alpha
  */
-public interface TriggerSupportCache extends BaseCache<StringIdKey, TriggerSupport> {
+public interface TriggerSupportCache extends BatchBaseCache<StringIdKey, TriggerSupport> {
 }